// frontend/src/api/generated/models/todoResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TodoApi.Gateway | v1
 * OpenAPI spec version: 1.0.0
 */

export interface TodoResponse {
  id: string;
  title: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  dueDate: string | null;
  isCompleted: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt: string | null;
}